feat: ✨ Rename actions to traces
