+++ src/cli.rs
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: GPL-3.0-or-later
 
-use crate::actions::Actions;
+use crate::traces::Traces;
 use clap::{Parser, Subcommand};
 
 #[derive(Debug, Parser)]
@@ -15,5 +15,5 @@ pub struct Arguments {
 
 #[derive(Debug, Subcommand)]
 pub enum Commands {
-    Actions(Actions),
+    Traces(Traces),
 }
