+++ src/traces/drivers.rs
@@ -2,27 +2,27 @@
 //
 // SPDX-License-Identifier: GPL-3.0-or-later
 
-use crate::actions::{processes::spoofing::Spoofing, Runnable};
+use crate::traces::{drivers::byovd::Byovd, Runnable};
 use clap::{Args, Subcommand};
 use std::error::Error;
 
-pub mod spoofing;
+pub mod byovd;
 
 #[derive(Debug, Args)]
-pub struct Processes {
+pub struct Drivers {
     #[clap(subcommand)]
     pub command: Commands,
 }
 
 #[derive(Debug, Subcommand)]
 pub enum Commands {
-    Spoofing(Spoofing),
+    Byovd(Byovd),
 }
 
-impl Runnable for Processes {
+impl Runnable for Drivers {
     fn run(&self) -> Result<(), Box<dyn Error>> {
         match &self.command {
-            Commands::Spoofing(spoofing) => spoofing as &dyn Runnable,
+            Commands::Byovd(byovd) => byovd as &dyn Runnable,
         }
         .run()
     }
