+++ src/traces/drivers/byovd.rs
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: GPL-3.0-or-later
 
-use crate::{actions::Runnable, windows::users::is_administrator};
+use crate::{traces::Runnable, windows::users::is_administrator};
 use clap::Parser;
 use std::{error::Error, path::PathBuf};
 use windows::{
